mcount: Calculate depth using frame pointer
With --flat option, it doesn't hook return address and we cannot know
when functions are returned.  But we can look how frame pointer
register value is changed and calculate depth based on it.  This will
help replay and graph commands to build call graph and report command
to estimate function execution time.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>