plthook: Resolve PLT addresses when --flat option is used
The PLT address is usually restored in the plthook_exit function since
the resolver function is called after the entry.  But with --flat
option there is no chance to do it since the exit path is not used.
So it needs to resolve PLT at init otherwise it cannot trace library
calls after the first invocation.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>