uftrace: Add --estimate-return option
The --estimate-return option is to run uftrace and a target process in
a safe way.  Normally uftrace runs a target process with hooking
return address of each function to get the timing of return
function (also to build a call graph).

However it's dangerous when the process wants to do something with the
execution stack as we changed it unexpectedly.  Using this option can
prevent many of crashes reported due to this problem.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>