filter: Add --hide option to filter functions not subtree
The --hide option will hide single depth of function and continue to
trace child functions.  This can be useful when user functions are
called from a library, but don't want to trace the library.

This option works both on record and replay time and 'hide' can be used
as a trigger just like other triggers such as filter.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>