plthook: Do not increase depth after a PLT function
In most cases, PLT symbols are leaf functions meaning that they don't
call other functions in the same module.  I added "in the same module"
since we know --nest-libcall can add other (PLT) functions in a PLT
call.  Also it can be a case when we trace functions in a library.

Exceptional cases would be re-defining/intercepting some well-known
library functions (like malloc) in the executable.  But this is not
common so I'd ignore them for better output in most cases.

This change will check previous record is from PLT in the same module
and finish the PLT call regardless of the frame pointer value if it's
the case.  It reduces unnecessary function call depth increment.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>