symbol: Update map file to have build-id info
The map files are written from libmcount by reading /proc/self/maps.
But we want to add build-id for each binary so rewrite the file after
the tracing is finished.  This is to prevent name clash in symbol
files as we only use base name for them.

FIXME: unit test 53 (task symbol) is failing.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>