filter: Add ARG_FMT_STRUCT type for aggregates
To handle struct/union/class types (esp. for pass-by-value) add
ARG_FMT_STRUCT.  They have 't' specifier followed by a size in
bytes.  (Well, maybe it's confusing as base types have bit size,
but I think it's intuitive having byte size for structs.)

After that it has an optional type name after a ':' (colon).

The uftrace replay will simply print them as "mystruct{...}".

Note that sometimes they can be passed via registers (note for the
plurals).  This makes hard to build and/or parse the argspec (for this
and next ones).  So they should have explicit location specifiers like
register names and stack offsets.  In case a struct is passed via
more than one registers, they should be separated by '+'.

For example, the following struct is 16 byte size, first field is
passed via integer registers and the second one is passed via a
floating-point register:

  arg1/t16:mixed%rdi+xmm0

If the following class is passed by stack, it looks like below:

  arg2/t56%stack+3

If type specifiers are omitted, it assumes it's passed by a pointer
and contents are allocated by a caller.  This is for backward
compatibility and actually happens a lot (like C++ classes).

  arg3/t40

This also allows an empty struct like below (not consuming a register):

  arg4/t0:empty

It's not intended to be used directly by (normal) users.  It will be
filled by using DWARF debug info.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>