filter: Add ARG_FMT_STRUCT type for aggregates
To handle struct/union types (esp. for pass-by-value) add
ARG_FMT_STRUCT.  The uftrace replay will print them as "{...}".

Note that they can contain some fields that can be passed via
registers.  This makes hard to build and/or parse the argspec.  It has
total size followed by an optional colon and register types.  For now
it only accepts 'i' for integer registers and 'f' for floating-point
registers.  They can be added up to 4 (according to x86_64 ABI).

For example, the following struct is 24 byte size, first two fields
are passed via integer registers and the last one is passed via a
floating-point register:

  arg1/t24:iif

If registers are omitted, the entire struct will be passed via stack:

  arg2/t40

This allows an empty struct like below:

  arg3/t0

It's not intended to be used by (normal) users.  It will be filled by
using DWARF debug info.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>