mcount: Start daemon in libmcount
This patch introduces a daemon inside libmcount, waiting for incoming
commands.

The daemon is started as a new thread when libmcount is set up. It is
listening to a socket at /var/run/user/%UID/uftrace/$PID.socket. It dies
when the target program finishes.

The daemon can be disabled with the --no-daemon option.