client: Add client command with base options
This patch adds the client command to uftrace. The client is intended to
be launched as a second uftrace instance, sending commands to a first
one running the daemon.

The daemon instance is selected through its PID with the -p <pid>
option.

The client can kill the daemon with the --kill option.