client: Filter and trigger options
This patch allows the client to add triggers to a program being traced.
It supports triggers (-T option), filters (-F and -N options) and caller
filters (-C option).

Arguments of the clients are sent to the daemon, which then modifies the
red black tree storing the function triggers inside libmcount.